% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banyan.R
\docType{package}
\name{banyan}
\alias{banyan-package}
\alias{banyan}
\title{mypackage: A package for fitting single and multilevel SBMs.}
\description{
This package fits Bayesian stochastic block models (SBMs)
}
\section{banyan functions}{

The banyan functions ...
}

\author{
\strong{Maintainer}: Kyeong Joo Jung \email{iamkj0829@gmail.com}

Authors:
\itemize{
  \item Carter Allen \email{carter.allen12@gmail.com} (\href{https://orcid.org/0000-0001-6937-7234}{ORCID})
  \item Dongjun Chung
}

}
