% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banyan_viz.R
\name{banyan_viz}
\alias{banyan_viz}
\title{Plot tissue architecture labels}
\usage{
banyan_viz(
  fit,
  pt.size = 1,
  interactive = FALSE,
  shade_uncertainty = FALSE,
  feature = NULL
)
}
\arguments{
\item{fit}{A list returned by fit_banyan()}

\item{pt.size}{The size of each cell spot point}

\item{interactive}{Logical parameter controlling static or interactive nature of plot}

\item{shade_uncertainty}{Logical parameter for shading of cell spots by posterior uncertainty. Must run get_scores() first.}

\item{feature}{A user-provided feature (e.g., gene of interest) to visualize over tissue spaces instead of sub-population labels.}
}
\value{
A ggplot object or shiny app window
}
\description{
This function allows you to plot (static or interactive) cell spot labels and uncertainty measures
}
\keyword{Bayesian}
\keyword{spatial}
\keyword{transcriptomics}
