% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_banyan.R
\name{fit_banyan}
\alias{fit_banyan}
\title{Fit Banyan community detection model for spatially resolved single cell data}
\usage{
fit_banyan(
  seurat_obj = NULL,
  exp = NULL,
  coords_df = NULL,
  z_init = NULL,
  K,
  n_pcs = 16,
  R = NULL,
  a0 = 2,
  b10 = 1,
  b20 = 1,
  n_iter = 1000,
  burn = 100,
  verbose = TRUE,
  s = 1.2
)
}
\arguments{
\item{seurat_obj}{A Seurat object with PCA reduction and spatial coordinates. If provided, the exp and coords arguments are ignored}

\item{exp}{A binary adjacency matrix encoding the gene expression network. Not used if seurat_obj is provided.}

\item{coords_df}{A matrix or data frame with rows as cells and 2 columns for coordinates. Rows should be ordered the same as in exp. Not used if seurat_obj is provided.}

\item{z_init}{An optional initialization for cluster indicators. Ignored if seurat_obj is provided.}

\item{K}{The number of sub-populations to infer}

\item{n_pcs}{The number of principal components to use from the Seurat object}

\item{R}{A length 2 vector of integers for the number of neighbors to use. 1st element corresponds to the number of neighbors in gene expression network and 2nd element for spatial.}

\item{a0}{Dirichlet prior parameter (shared across all communities)}

\item{b10}{Beta prior number of connections}

\item{b20}{Beta prior number on non-edges}

\item{n_iter}{The number of total MCMC iterations to run.}

\item{burn}{The number of MCMC iterations to discard as burn-in. A total of n_iter - burn iterations will be saved.}

\item{verbose}{Whether or not to print cluster allocations at each iteration}

\item{s}{Louvain resolution parameter}
}
\value{
A list of MCMC samples, including the MAP estimate of cluster indicators (z)
}
\description{
This function allows you to fit the Bayesian multi-layer SBM for integration of spatial and gene expression data for identifying cell sub-populations
}
\keyword{Bayesian}
\keyword{Gibbs}
\keyword{MLSBM}
\keyword{SBM}
\keyword{expression}
\keyword{gene}
\keyword{networks}
\keyword{spatial}
