% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ACC.R
\name{get_ACC}
\alias{get_ACC}
\title{Get ACC parameters}
\usage{
get_ACC(z, AL, b10, b20)
}
\arguments{
\item{z}{User-defined cell spot labels}

\item{AL}{List of adjacency matrices}

\item{b10}{beta prior parameter 1}

\item{b20}{beta prior parameter 2}
}
\value{
an adjacency matrix
}
\description{
Conduct ACC based on user-defined tissue architecture labels
}
