% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores}
\alias{get_scores}
\title{Calculate continuous uncertainty scores}
\usage{
get_scores(fit)
}
\arguments{
\item{fit}{A list returned by fit_banyan()}
}
\value{
A list with populated entries C_scores (N x K matrix for cell type propensities) and U_scores (N x 1 vector of uncertainty scores)
}
\description{
This function allows you to augment the discrete cell type assignments with continuous propensity and uncertainty scores
}
\keyword{Bayesian}
\keyword{Gibbs}
\keyword{MLSBM}
\keyword{SBM}
\keyword{expression}
\keyword{gene}
\keyword{networks}
\keyword{spatial}
