% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_adjacency_matrix.R
\name{plot_adjacency_matrix}
\alias{plot_adjacency_matrix}
\title{Plot cell-cell adjacency matrix sorted by labels from fit_banyan()}
\usage{
plot_adjacency_matrix(fit, layer = "expression")
}
\arguments{
\item{fit}{A list returned by fit_banyan(). Must have slots named "A" and "z".}

\item{layer}{Either "expression" (default) or "spatial" for which adjacency network to use}
}
\value{
A ggplot object
}
\description{
This function allows you to visualize the inferred adjacency matrix after running fit_banyan()
}
\keyword{Bayesian}
\keyword{Gibbs}
\keyword{MLSBM}
\keyword{SBM}
\keyword{expression}
\keyword{gene}
\keyword{networks}
\keyword{spatial}
