% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connectivity_intervals.R
\name{plot_connectivity_intervals}
\alias{plot_connectivity_intervals}
\title{Plot community structure of cell sub-populations as posterior credible intervals}
\usage{
plot_connectivity_intervals(fit)
}
\arguments{
\item{fit}{A list returned by fit_banyan().}
}
\value{
A ggplot object
}
\description{
This function allows you to visualize the community structure of cell sub-populations in posterior credible interval format via the connectivity parameters of the BANYAN model
}
\keyword{Bayesian}
\keyword{Gibbs}
\keyword{MLSBM}
\keyword{SBM}
\keyword{expression}
\keyword{gene}
\keyword{networks}
\keyword{spatial}
