% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_connectivity_network.R
\name{plot_connectivity_network}
\alias{plot_connectivity_network}
\title{Plot community structure parameters as a K x K network}
\usage{
plot_connectivity_network(fit)
}
\arguments{
\item{fit}{A list returned by fit_banyan()}
}
\value{
A ggplot object
}
\description{
This function allows you to visualize the inferred community structure as a community-community connectivity network
}
\keyword{Bayesian}
\keyword{Gibbs}
\keyword{MLSBM}
\keyword{SBM}
\keyword{expression}
\keyword{gene}
\keyword{networks}
\keyword{spatial}
