% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_labels.R
\name{plot_labels}
\alias{plot_labels}
\title{Plot tissue labels from fit_banyan()}
\usage{
plot_labels(fit)
}
\arguments{
\item{fit}{A list returned by fit_banyan(). Must have slots named "coords" and "z".}
}
\value{
A ggplot object
}
\description{
This function allows you to visualize the inferred cell sub-populations after running fit_banyan()
}
\keyword{Bayesian}
\keyword{Gibbs}
\keyword{MLSBM}
\keyword{SBM}
\keyword{expression}
\keyword{gene}
\keyword{networks}
\keyword{spatial}
