% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_propensity.R
\name{plot_propensity}
\alias{plot_propensity}
\title{Plot continuous phenotype scores from fit_banyan()}
\usage{
plot_propensity(fit, k = 1)
}
\arguments{
\item{fit}{A list returned by fit_banyan(). Must have slots named "coords" and "C_scores".}

\item{k}{Which cell sub-population to compute propensity towards}
}
\value{
A ggplot object
}
\description{
This function allows you to visualize continuous propensity of each cell spot towards each sub-population (e.g., continuous phenoptypes) after running fit_banyan() and get_scores()
}
\keyword{Bayesian}
\keyword{Gibbs}
\keyword{MLSBM}
\keyword{SBM}
\keyword{expression}
\keyword{gene}
\keyword{networks}
\keyword{spatial}
